Move pkg/infra/ent/schema to pkg/infra/schema
