+++ pkg/infra/ent/runtime/runtime.go
@@ -3,8 +3,3 @@
 package runtime
 
 // The schema-stitching logic is generated in github.com/m-mizutani/alertchain/pkg/infra/ent/runtime.go
-
-const (
-	Version = "v0.9.0"                                          // Version of ent codegen.
-	Sum     = "h1:2S1zfpMMW6p+wctj6kcYUprNPNjLWFW06T5MdyAfmWc=" // Sum of ent codegen.
-)
